+++ paging/src/commonMain/kotlin/ru/astrainteractive/klibs/paging/IntPagerCollector.kt
@@ -1,7 +1,7 @@
 package ru.astrainteractive.klibs.paging
 
 import ru.astrainteractive.klibs.paging.data.PagedListDataSource
-import ru.astrainteractive.klibs.paging.descriptor.IntPageDescriptor
+import ru.astrainteractive.klibs.paging.context.IntPageContext
 import ru.astrainteractive.klibs.paging.state.PagingState
 
 class IntPagerCollector<T>(
@@ -10,7 +10,7 @@ class IntPagerCollector<T>(
     private val pager: PagedListDataSource<T, Int>,
 ) : PagingCollector<T, Int> by DefaultPagingCollector(
     initialPagingState = PagingState(
-        pageDescriptor = IntPageDescriptor(value = initialPage),
+        pageContext = IntPageContext(value = initialPage),
         pageSizeAtLeast = pageSize,
         isLastPage = false,
         isLoading = false,
