+++ paging/src/commonMain/kotlin/ru/astrainteractive/klibs/paging/state/PagingState.kt
@@ -1,15 +1,15 @@
 package ru.astrainteractive.klibs.paging.state
 
-import ru.astrainteractive.klibs.paging.descriptor.PageDescriptor
+import ru.astrainteractive.klibs.paging.context.PageContext
 
 /**
  * Core paging state
  * [T] is the type of page - can be int; string whatever
  *
- * @see PageDescriptor
+ * @see PageContext
  */
 data class PagingState<T : Any>(
-    val pageDescriptor: PageDescriptor<T>,
+    val pageContext: PageContext<T>,
     val pageSizeAtLeast: Int,
     val isLastPage: Boolean,
     val isLoading: Boolean,
